requirements.txt
Jinja2==2.7.2
dnspython==1.11.1
requests==1.1.0
